feat: implement security middleware and fix registration/onboarding validation
- Migrate registration API to use secureEndpoint with proper validation
- Fix registration password validation to match backend requirements
- Update onboarding completion schema to not require firstName/lastName
- Fix habit reminderTime format validation (24-hour format)
- Preserve existing user name during onboarding completion
- Add comprehensive security middleware with rate limiting and sanitization
- Remove debugging files and clean up unused endpoints
