+++ src/app/globals.css
@@ -202,7 +202,7 @@
     --mv-form-input-bg: #ffffff;
     --mv-form-input-text: #0f172a;
     --mv-form-input-placeholder: #64748b;
-    --mv-form-input-border: #cbd5e1;
+    --mv-form-input-border: #9ca3af;
 
     /* Dashboard specific colors */
     --mv-dashboard-bg: #f8fafc; /* Professional gray-50 equivalent */
@@ -383,7 +383,7 @@
 
   /* Light theme form inputs */
   .light .form-input {
-    @apply bg-white border border-gray-300 text-gray-900 placeholder-gray-500 focus:bg-white focus:border-blue-500;
+    @apply bg-white border border-gray-400 text-gray-900 placeholder-gray-500 focus:outline-none focus:bg-white focus:border-blue-500 focus:ring-2 focus:ring-blue-500/20 transition-all duration-200;
   }
 
   .form-checkbox {
